requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional - for mobile export
# coremltools>=7.0  # iOS export
# onnx2tf>=1.17.0   # TFLite export
